fix: surface native export warnings
